Move contrib.render functionality to infer.inspect (#1447)
