+++ docs/source/utilities.rst
@@ -101,14 +101,8 @@ Model inspection
     :members:
     :member-order: bysource
 
-.. autofunction:: numpyro.contrib.render.get_model_relations
-
 Visualization Utilities
 =======================
-
-render_model
-------------
-.. autofunction:: numpyro.contrib.render.render_model
 
 Trace inspection
 ----------------
